Braze-Script/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
PyMySQL==1.0.3
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
yarl==1.9.2
